moved exception to exceptions file
